refactor: move getImagesAction from `utils` to `actions` folder
