+++ src/actions/getImagesAction.ts
